# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cloud Storage API using Rclone with multiple providers support, authentication, and video streaming capabilities.",
        "title": "RcloneStorage API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "RcloneStorage API Support",
            "email": "support@rclonestorage.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stream/{id}": {
            "get": {
                "description": "Stream video file with range support for progressive loading",
                "produces": [
                    "video/*"
                ],
                "tags": [
                    "streaming"
                ],
                "summary": "Stream video file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Range header for partial content",
                        "name": "Range",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video stream",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "206": {
                        "description": "Partial content",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/stream/{id}/info": {
            "get": {
                "description": "Get video stream information including duration, bitrate, resolution",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming"
                ],
                "summary": "Get stream info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stream information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}